refactor: extracted parser into class
