+++ src/main/java/TaskList.java
@@ -1,0 +1,2 @@
+package PACKAGE_NAME;public class TaskList {
+}
